# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a manpower management system API.",
        "title": "Manpower API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/user/requests": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ดึงรายการ manpower requests ตามสิทธิ์ (Admin/Approve เห็นทั้งหมด, User เห็นเฉพาะกอง/แผนกตัวเอง)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Requests"
                ],
                "summary": "Get all manpower requests",
                "responses": {
                    "200": {
                        "description": "data: []ManpowerRequest",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/user/requests/submit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ผู้ใช้สร้างคำขอและส่งเข้ากระบวนการอนุมัติทันที (origin_status=SUBMITTED)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Requests"
                ],
                "summary": "Create \u0026 submit manpower request",
                "parameters": [
                    {
                        "description": "payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateManpowerRequestInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateSubmitResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.CreateManpowerRequestInput": {
            "type": "object",
            "required": [
                "contract_type_id",
                "employment_type_id",
                "num_required",
                "reason_id",
                "required_position_name"
            ],
            "properties": {
                "contract_type_id": {
                    "type": "integer"
                },
                "education_level_id": {
                    "type": "integer"
                },
                "employment_type_id": {
                    "type": "integer"
                },
                "experience_id": {
                    "type": "integer"
                },
                "gender_id": {
                    "type": "integer"
                },
                "max_age": {
                    "type": "integer"
                },
                "min_age": {
                    "type": "integer"
                },
                "nationality_id": {
                    "type": "integer"
                },
                "num_required": {
                    "type": "integer"
                },
                "reason_id": {
                    "type": "integer"
                },
                "required_position_name": {
                    "type": "string"
                },
                "special_qualifications": {
                    "type": "string"
                },
                "target_hire_date": {
                    "description": "รูปแบบ YYYY-MM-DD",
                    "type": "string"
                }
            }
        },
        "handlers.CreateSubmitResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2025-10-12T15:12:00Z"
                },
                "created_by": {
                    "type": "string",
                    "example": "E101"
                },
                "doc_number": {
                    "type": "string",
                    "example": "PQ25100001"
                },
                "hr_status": {
                    "type": "string",
                    "example": "NONE"
                },
                "message": {
                    "type": "string",
                    "example": "Submitted successfully"
                },
                "origin_status": {
                    "type": "string",
                    "example": "SUBMITTED"
                },
                "overall_status": {
                    "type": "string",
                    "example": "IN_PROGRESS"
                },
                "request_id": {
                    "type": "integer",
                    "example": 123
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}